split authorization components into separate files
